Updated adding of metals, including updating strength to latest desisim values
